Move tests to test_ModelUtils.py
